+++ Source/Core/FontEngineDefault/FontEngineInterfaceDefault.cpp
@@ -101,10 +101,10 @@ int FontEngineInterfaceDefault::GetStringWidth(FontFaceHandle handle, const Stri
 }
 
 int FontEngineInterfaceDefault::GenerateString(FontFaceHandle handle, FontEffectsHandle font_effects_handle, const String& string,
-	const Vector2f& position, const Colourb& colour, GeometryList& geometry)
+	const Vector2f& position, const Colourb& colour, float opacity, GeometryList& geometry)
 {
 	auto handle_default = reinterpret_cast<FontFaceHandleDefault *>(handle);
-	return handle_default->GenerateString(geometry, string, position, colour, (int)font_effects_handle);
+	return handle_default->GenerateString(geometry, string, position, colour, opacity, (int)font_effects_handle);
 }
 
 int FontEngineInterfaceDefault::GetVersion(FontFaceHandle handle)