+++ changelog.md
@@ -10,9 +10,26 @@
 
 ## RmlUi 4.4 (WIP)
 
-### Layout improvements
+### Improvements
+
+- Release memory pools on `Rml::Shutdown`, or manually through the core API. [#263](https://github.com/mikke89/RmlUi/issues/263) [#265](https://github.com/mikke89/RmlUi/pull/265) (thanks @jack9267)
+- Fix classes not always copied over to a cloned element. [#264](https://github.com/mikke89/RmlUi/issues/264)
+- `select` element: Fix clipping on select box.
+- The `opacity` property is now also applied to font effects. [#270](https://github.com/mikke89/RmlUi/issues/270)
+- Drag clones are now positioned correctly when their ancestors use transforms. [#269](https://github.com/mikke89/RmlUi/issues/269)
+- Drag clones no longer inherit backgrounds and decoration from the cloned element's document body.
+
+### Layout
 
 - Fix offsets of relatively positioned elements with percentage positioning. [#262](https://github.com/mikke89/RmlUi/issues/262)
+
+### Samples
+
+- New sample for integration with SDL2's native renderer. [#252](https://github.com/mikke89/RmlUi/pull/252) (thanks @1bsyl)
+
+### Breaking changes
+
+- `FontEngineInterface::GenerateString` now takes a new argument, `opacity`.
 
 
 ## RmlUi 4.3